requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
dnspython==1.16.0
future==0.17.1
idna==2.8
pycparser==2.19
pymongo==3.7.2
python-telegram-bot==11.1.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
